Feature/WIOA-210 adjust node creation

* WIOA-210: Added field_unique_id_reference field to state plan content in order to know which piece term and input from state it belongs to.

* WIOA-210: Updated auto titles to be under 255 characters and show more info on state content nodes.

* WIOA-210: Added permissions for sys admin.

* WIOA-210: Removed edit permission for plan year entities for sys admin.

* WIOA-210: Added plan year overview block at the top of the page to edit plans. Added the ability to create missing content (nodes used to answer questions).

* WIOA-210: Removed missing validator.

Approved-by: Gerardo Maldonado <gerardo.maldonado@bixal.com>

+++ web/modules/custom/sp_create/src/UpdatePlanYearConfigService.php
@@ -112,8 +112,10 @@ class UpdatePlanYearConfigService {
     // Remove all section based information.
     /** @var \Drupal\node\Entity\Node $state_plans_year */
     foreach ($plan_year->getSections() as $section) {
-      // Remove content and state plan year section nodes for this section.
-      $this->updatePlanYearContentService->removeSectionContent($plan_year_id, $section->id());
+      // Remove state plan year content nodes for this section.
+      $this->updatePlanYearContentService->removeStatePlanYearContentBySection($plan_year_id, $section->id());
+      // Remove all section nodes in this section.
+      $this->updatePlanYearContentService->removeStatePlanYearSection($plan_year_id, $section->id());
       // Remove all terms in the plan year section vocabulary.
       $this->updatePlanYearContentService->removeSectionHierarchy($plan_year_id, $section->id());
       // Remove the plan year section vocabulary.
@@ -127,9 +129,12 @@ class UpdatePlanYearConfigService {
       }
     }
     // Get the state plans year node for this plan year and remove it.
+    // This content might not be created yet.
     $state_plans_year_nid = $this->nodeService->getStatePlansYearByPlanYear($plan_year_id);
-    $state_plans_year = $node_storage->load($state_plans_year_nid);
-    $state_plans_year->delete();
+    if (!empty($state_plan_year_nid)) {
+      $state_plans_year = $node_storage->load($state_plans_year_nid);
+      $state_plans_year->delete();
+    }
   }
 
   /**
